refactor: extract InfoSection from MyPage
