+++ src/pages/MyPage/components/Mypage/index.ts
@@ -1,0 +1,2 @@
+export { default as InfoSection } from './InfoSection';
+export { default as ProfileCard } from './ProfileCard';
